+++ SystemBackEnd/ServiceModels/AdjustItemDetailswithItemName.cs
@@ -12,7 +12,7 @@ namespace SystemBackEnd.ServiceModels
         public Guid ReferenceHeaderId { get; set; }
         public Guid ItemId { get; set; }
         public string? Reason { get; set; }
-        public int RemovedQuantity { get; set; }
+        public decimal RemovedQuantity { get; set; }
         public string ItemName { get; set; }
     }
 }
