+++ SystemBackEnd/ServiceModels/AdjustmentItemFullInformation.cs
@@ -19,7 +19,7 @@ namespace SystemBackEnd.ServiceModels
     {
         public Guid ItemId { get; set; }
         public string Reason { get; set; }
-        public int ItemQuantity { get; set; }
+        public decimal ItemQuantity { get; set; }
 
     }
 }
